Completed module

+++ WishList/Controllers/ItemController.cs
@@ -1,21 +1,28 @@
 ﻿using System.Linq;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using WishList.Data;
+using WishList.Models;
 
 namespace WishList.Controllers
 {
+    [Authorize]
     public class ItemController : Controller
     {
         private readonly ApplicationDbContext _context;
+        private readonly UserManager<ApplicationUser> _userManager;
 
-        public ItemController(ApplicationDbContext context)
+        public ItemController(ApplicationDbContext context, UserManager<ApplicationUser> userManager)
         {
             _context = context;
+            _userManager = userManager;
         }
 
         public IActionResult Index()
         {
-            var model = _context.Items.ToList();
+            var user = _userManager.GetUserAsync(HttpContext.User);
+            var model = _context.Items.Where(x => x.User.Id == user.Result.Id).ToList();
 
             return View("Index", model);
         }
@@ -29,14 +36,21 @@ namespace WishList.Controllers
         [HttpPost]
         public IActionResult Create(Models.Item item)
         {
-            _context.Items.Add(item);
+            var user = _userManager.GetUserAsync(HttpContext.User);
+            item.User.Id = user.Result.Id;
+            _context.Items.Add(item); 
             _context.SaveChanges();
             return RedirectToAction("Index");
         }
 
         public IActionResult Delete(int id)
         {
+            var user = _userManager.GetUserAsync(HttpContext.User);
             var item = _context.Items.FirstOrDefault(e => e.Id == id);
+            if(item.User.Id != user.Result.Id)
+            {
+                return Unauthorized();
+            }
             _context.Items.Remove(item);
             _context.SaveChanges();
             return RedirectToAction("Index");
